beforeDevWork/Advanced RAG Implementation/requirements.txt
numpy>=1.26.0
chromadb>=0.5.3
pandas>=2.0.0
pypdf>=5.1.0
openai>=1.60.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.29
langchain-text-splitters>=0.0.1
umap-learn>=0.5.0
matplotlib>=3.9.0
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.0.0
streamlit>=1.29.0
python-docx>=0.8.11
html2text>=2020.1.16
jieba>=0.42.1
zhconv>=1.4.3
chardet>=4.0.0